refactor(service): 파일 저장 서비스 -> 도메인으로 이동
